% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_wide.R
\name{make_wide}
\alias{make_wide}
\title{Wide format for contaminants in marine biota database.}
\usage{
make_wide(
  ContaminantsDB,
  group,
  bio = c("year", "date", "species", "species_EN", "class", "number_individuals", "sex",
    "total_length", "age", "weight", "shell_thickness", "d13C", "d15N",
    "C_dw_percentage", "N_dw_percentage", "station_name", "latitude", "longitude",
    "HELCOM_basin", "ICES_basin", "SMNH_basin"),
  Averages = TRUE
)
}
\arguments{
\item{ContaminantsDB}{data table of contaminants in marine biota presented in a long format.}

\item{group}{the contaminants substance groups. Choose from:  "all", "Metals","PFAS","PCBs", "Dioxins and furans" , "Pesticides","BFRs", "Solvents-phenols","PAH". "Organotin compounds" .}

\item{bio}{choose from: "Year", "Date", "Species", "English_name",  "Class","Number_individual","Sex", "Total_length","Age", "Weight","Shell_thickness" ,  "d13C", "d15N","C_dw_percentage", "N_dw_percentage","Station_name", "Latitude", "Longitude", "Helcom_basin", "Ices_basin","SMNH_basin".}

\item{Averages}{add the average of fat percentage and dry weight percentage.}
}
\value{
database in wide format with contaminants in separate columns.
}
\description{
Wide format for contaminants in marine biota database.
}
\examples{
make_wide(ContaminantsDB, "Metals") |>  glimpse() |> View()
make_wide(ContaminantsDB, "BFRs") |> write_csv2(file = "BFRs.csv")

}
