% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMNHcontaminants-package.R
\docType{package}
\name{SMNHcontaminants-package}
\alias{SMNHcontaminants}
\alias{SMNHcontaminants-package}
\title{SMNHcontaminants: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
